+++ bs-config.json
@@ -1,0 +1,5 @@
+{
+    "port": 3000,
+    "files": ["./vanilla/**/*.{html,htm,css,js}"],
+    "server": { "baseDir": "./vanilla" }
+  }
