--- src/scss/_variable.scss
@@ -1,3 +1,0 @@
-// Custom variables overrides
-$primary: #0074d9;
-$danger: #ff4136;
