--- src/scss/main.scss
@@ -1,4 +1,0 @@
-// Your custom styles go here
-@import "variable";
-
-@import "../../node_modules/bootstrap/scss/bootstrap";
