+++ vanilla/data/books.json
@@ -1,0 +1,274 @@
+{
+  "books": [
+    {
+      "title": "The Catcher in the Rye",
+      "author": "J.D. Salinger",
+      "year_published": 1951
+    },
+    {
+      "title": "Pride and Prejudice",
+      "author": "Jane Austen",
+      "year_published": 1813
+    },
+    {
+      "title": "Moby-Dick",
+      "author": "Herman Melville",
+      "year_published": 1851
+    },
+    {
+      "title": "One Hundred Years of Solitude",
+      "author": "Gabriel Garcia Marquez",
+      "year_published": 1967
+    },
+    {
+      "title": "War and Peace",
+      "author": "Leo Tolstoy",
+      "year_published": 1869
+    },
+    {
+      "title": "The Grapes of Wrath",
+      "author": "John Steinbeck",
+      "year_published": 1939
+    },
+    {
+      "title": "Jane Eyre",
+      "author": "Charlotte Bronte",
+      "year_published": 1847
+    },
+    {
+      "title": "The Lord of the Rings",
+      "author": "J.R.R. Tolkien",
+      "year_published": 1954
+    },
+    {
+      "title": "Wuthering Heights",
+      "author": "Emily Bronte",
+      "year_published": 1847
+    },
+    {
+      "title": "The Picture of Dorian Gray",
+      "author": "Oscar Wilde",
+      "year_published": 1890
+    },
+    {
+      "title": "The Odyssey",
+      "author": "Homer",
+      "year_published": -800
+    },
+    {
+      "title": "Frankenstein",
+      "author": "Mary Shelley",
+      "year_published": 1818
+    },
+    {
+      "title": "Brave New World",
+      "author": "Aldous Huxley",
+      "year_published": 1932
+    },
+    {
+      "title": "The Canterbury Tales",
+      "author": "Geoffrey Chaucer",
+      "year_published": 1387
+    },
+    {
+      "title": "The Adventures of Huckleberry Finn",
+      "author": "Mark Twain",
+      "year_published": 1884
+    },
+    {
+      "title": "Crime and Punishment",
+      "author": "Fyodor Dostoevsky",
+      "year_published": 1866
+    },
+    {
+      "title": "Don Quixote",
+      "author": "Miguel de Cervantes",
+      "year_published": 1605
+    },
+    {
+      "title": "The Scarlet Letter",
+      "author": "Nathaniel Hawthorne",
+      "year_published": 1850
+    },
+    {
+      "title": "A Tale of Two Cities",
+      "author": "Charles Dickens",
+      "year_published": 1859
+    },
+    {
+      "title": "The Count of Monte Cristo",
+      "author": "Alexandre Dumas",
+      "year_published": 1844
+    },
+    {
+      "title": "Great Expectations",
+      "author": "Charles Dickens",
+      "year_published": 1861
+    },
+    {
+      "title": "Anna Karenina",
+      "author": "Leo Tolstoy",
+      "year_published": 1877
+    },
+    {
+      "title": "The Brothers Karamazov",
+      "author": "Fyodor Dostoevsky",
+      "year_published": 1880
+    },
+    {
+      "title": "The Divine Comedy",
+      "author": "Dante Alighieri",
+      "year_published": 1320
+    },
+    {
+      "title": "Les Misérables",
+      "author": "Victor Hugo",
+      "year_published": 1862
+    },
+    {
+      "title": "The Wind in the Willows",
+      "author": "Kenneth Grahame",
+      "year_published": 1908
+    },
+    {
+      "title": "The Sun Also Rises",
+      "author": "Ernest Hemingway",
+      "year_published": 1926
+    },
+    {
+      "title": "Siddhartha",
+      "author": "Hermann Hesse",
+      "year_published": 1922
+    },
+    {
+      "title": "The Stranger",
+      "author": "Albert Camus",
+      "year_published": 1942
+    },
+    {
+      "title": "The Bell Jar",
+      "author": "Sylvia Plath",
+      "year_published": 1963
+    },
+    {
+      "title": "The Old Man and the Sea",
+      "author": "Ernest Hemingway",
+      "year_published": 1952
+    },
+    {
+      "title": "Mrs. Dalloway",
+      "author": "Virginia Woolf",
+      "year_published": 1925
+    },
+    {
+      "title": "Catch-22",
+      "author": "Joseph Heller",
+      "year_published": 1961
+    },
+    {
+      "title": "A Clockwork Orange",
+      "author": "Anthony Burgess",
+      "year_published": 1962
+    },
+    {
+      "title": "Heart of Darkness",
+      "author": "Joseph Conrad",
+      "year_published": 1899
+    },
+    {
+      "title": "The Jungle",
+      "author": "Upton Sinclair",
+      "year_published": 1906
+    },
+    {
+      "title": "Slaughterhouse-Five",
+      "author": "Kurt Vonnegut",
+      "year_published": 1969
+    },
+    {
+      "title": "The Call of the Wild",
+      "author": "Jack London",
+      "year_published": 1903
+    },
+    {
+      "title": "Lord of the Flies",
+      "author": "William Golding",
+      "year_published": 1954
+    },
+    {
+      "title": "The Color Purple",
+      "author": "Alice Walker",
+      "year_published": 1982
+    },
+    {
+      "title": "Gulliver's Travels",
+      "author": "Jonathan Swift",
+      "year_published": 1726
+    },
+    {
+      "title": "The Hobbit",
+      "author": "J.R.R. Tolkien",
+      "year_published": 1937
+    },
+    {
+      "title": "The Metamorphosis",
+      "author": "Franz Kafka",
+      "year_published": 1915
+    },
+    {
+      "title": "The Sound and the Fury",
+      "author": "William Faulkner",
+      "year_published": 1929
+    },
+    {
+      "title": "The Iliad",
+      "author": "Homer",
+      "year_published": -800
+    },
+    {
+      "title": "The War of the Worlds",
+      "author": "H.G. Wells",
+      "year_published": 1898
+    },
+    {
+      "title": "Beloved",
+      "author": "Toni Morrison",
+      "year_published": 1987
+    },
+    {
+      "title": "The Martian Chronicles",
+      "author": "Ray Bradbury",
+      "year_published": 1950
+    },
+    {
+      "title": "Their Eyes Were Watching God",
+      "author": "Zora Neale Hurston",
+      "year_published": 1937
+    },
+    {
+      "title": "The Fountainhead",
+      "author": "Ayn Rand",
+      "year_published": 1943
+    },
+    {
+      "title": "Walden",
+      "author": "Henry David Thoreau",
+      "year_published": 1854
+    },
+    {
+      "title": "Tess of the d'Urbervilles",
+      "author": "Thomas Hardy",
+      "year_published": 1891
+    },
+    {
+      "title": "The Road",
+      "author": "Cormac McCarthy",
+      "year_published": 2006
+    },
+    {
+      "title": "The Handmaid's Tale",
+      "author": "Margaret Atwood",
+      "year_published": 1985
+    }
+  ]
+}
