+++ vanilla/index.html
@@ -1,0 +1,60 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>FED Assessment</title>
+    <link rel="stylesheet" href="dist/css/main.css">
+</head>
+<body>
+
+    <!-- Navigation -->
+    <nav class="navbar navbar-expand-lg navbar-light bg-light">
+        <div class="container">
+            <a class="navbar-brand" href="#">FED Assessment</a>
+            <button class="navbar-toggler" type="button" data-bs-toggle="collapse" data-bs-target="#navbarNav" aria-controls="navbarNav" aria-expanded="false" aria-label="Toggle navigation">
+                <span class="navbar-toggler-icon"></span>
+            </button>
+            <div class="collapse navbar-collapse" id="navbarNav">
+                <ul class="navbar-nav ml-auto">
+                    <li class="nav-item active">
+                        <a class="nav-link" href="#">Home</a>
+                    </li>
+                    <li class="nav-item">
+                        <a class="nav-link" href="#">About</a>
+                    </li>
+                    <li class="nav-item">
+                        <a class="nav-link" href="#">Contact</a>
+                    </li>
+                </ul>
+            </div>
+        </div>
+    </nav>
+
+
+    <!-- Header -->
+    <header class="jumbotron text-center">
+        <h1 class="display-4">Welcome to FED Assessment</h1>
+        <p class="lead">Lorem ipsum, dolor sit amet consectetur adipisicing elit. Ea repellat aliquam quas aliquid, doloribus deleniti minima perferendis rem laborum ullam voluptatibus odio impedit consequatur rerum nobis libero quos enim. Eius.</p>
+    </header>
+
+    <!-- Main Content - Book listing -->
+    <main>
+        <div class="container">
+            <h1>Book Listing</h1>
+            <input type="text" id="searchInput" class="form-control mb-3" placeholder="Search by title or author">
+            <div class="row" id="bookList"></div>
+        </div>
+    </main>
+
+    <div class="py-5"></div>
+
+    <!-- Footer -->
+    <footer class="container-fluid bg-dark text-light text-center py-3">
+        <p>&copy; 2024 FED Assessment. All rights reserved.</p>
+    </footer>
+
+    <script src="dist/js/main.js"></script>
+
+</body>
+</html>
