+++ vanilla/src/scss/_variable.scss
@@ -1,0 +1,3 @@
+// Custom variables overrides
+$primary: #0074d9;
+$danger: #ff4136;
